INDY-1911: skip test_view_change_with_different_ic
